requirements.txt
asgiref==3.5.2
Django==4.1.4
django-auto-prefetch==1.4.0
django-ckeditor==6.5.1
django-environ==0.9.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-resized==1.0.2
django-taggit==3.1.0
django_debug_toolbar==3.8.1
Pillow==9.3.0
python-utils==3.4.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
